#374: refactor 助理系统模块化重构
