+++ issue-comment-375.md
@@ -1,0 +1,57 @@
+## Phase 3: 模块拆分完成报告
+
+### 📊 重构概览
+
+将 `server/services/assistant-manager.js` (2398 行) 拆分为 7 个独立模块，提升代码可维护性和可测试性。
+
+### 📁 新模块结构
+
+```
+server/services/assistant/
+├── index.js              # 入口文件，统一导出
+├── manager.js            # 核心协调器 (~400 行)
+├── config-repository.js  # 配置管理 (~150 行)
+├── request-repository.js # 请求管理 (~280 行)
+├── executor.js           # 执行引擎 (~350 行)
+├── tool-integration.js   # 工具集成 (~180 行)
+├── vision-processor.js   # 视觉处理 (~350 行)
+└── expert-notifier.js    # 专家通知 (~450 行)
+```
+
+### 📋 模块职责
+
+| 模块 | 职责 | 主要函数 |
+|------|------|----------|
+| `config-repository.js` | 助理配置 CRUD | `refreshAssistantsCache`, `getAssistant`, `createAssistant`, `updateAssistant` |
+| `request-repository.js` | 请求生命周期 | `createRequest`, `getStatus`, `listRequests`, `archiveRequest`, `deleteRequest` |
+| `executor.js` | 多模式执行 | `executeAssistant`, `executeDirect`, `executeLLM`, `executeLLMWithTools` |
+| `tool-integration.js` | 工具定义与执行 | `getAssistantTools`, `getInheritedToolDefinitions`, `executeInheritedTool` |
+| `vision-processor.js` | 图片处理 | `extractImageInput`, `executeVisionWithInput`, `readImageFile` |
+| `expert-notifier.js` | SSE 推送与专家响应 | `pushSSENotification`, `notifyExpertResult`, `triggerExpertResponse` |
+| `manager.js` | 协调器 | `AssistantManager` 类，整合所有子模块 |
+
+### 🔄 导入路径更新
+
+| 文件 | 旧路径 | 新路径 |
+|------|--------|--------|
+| `server/controllers/assistant.controller.js` | `../services/assistant-manager.js` | `../services/assistant/index.js` |
+| `server/index.js` | `./services/assistant-manager.js` | `./services/assistant/index.js` |
+| `lib/tool-manager.js` | `../server/services/assistant-manager.js` | `../server/services/assistant/index.js` |
+
+### ✅ 验证结果
+
+```bash
+node -e "import('./server/services/assistant/index.js').then(() => console.log('Import successful')).catch(e => console.error('Import failed:', e.message))"
+# Output: Import successful
+```
+
+### 📈 收益
+
+1. **可维护性提升**：每个模块职责单一，代码行数控制在 400 行以内
+2. **可测试性提升**：各模块可独立测试
+3. **可复用性提升**：子模块函数可被其他服务直接调用
+4. **代码复用**：继续使用 Phase 1/2 创建的 `SimpleLLMClient` 和 `ToolCallingExecutor`
+
+### 🔗 相关 Issue
+
+- #374 - 代码复用分析与重构
