+++ lib/tool-manager.js
@@ -13,7 +13,7 @@
 
 import SkillLoader from './skill-loader.js';
 import logger from './logger.js';
-import { getAssistantManager } from '../server/services/assistant-manager.js';
+import { getAssistantManager } from '../server/services/assistant/index.js';
 import path from 'path';
 import { fileURLToPath } from 'url';
